+++ nowcasting_dataset/dataset/dataset.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from numbers import Number
-from typing import List, Tuple, Iterable, Callable, Optional
+from typing import List, Tuple, Iterable, Callable
 
 import nowcasting_dataset.data_sources.constants
 from nowcasting_dataset import data_sources
@@ -18,10 +18,8 @@ import torch
 
 from nowcasting_dataset.cloud.gcp import gcp_download_to_local
 from nowcasting_dataset.cloud.aws import aws_download_to_local
-from nowcasting_dataset.data_sources.constants import *
 
 from nowcasting_dataset.data_sources.satellite_data_source import SAT_VARIABLE_NAMES
-from nowcasting_dataset.example import Example
 
 SAT_MEAN = xr.DataArray(
     data=[
@@ -38,7 +36,7 @@ SAT_STD = xr.DataArray(
     dims=['sat_variable'],
     coords={'sat_variable': list(SAT_VARIABLE_NAMES)}).astype(np.float32)
 
-_LOG = logging.getLogger('nowcasting_dataset')
+_LOG = logging.getLogger(__name__)
 
 
 class NetCDFDataset(torch.utils.data.Dataset):
@@ -289,100 +287,3 @@ def worker_init_fn(worker_id):
         dataset_obj.per_worker_init(worker_info.id)
 
 
-def batch_to_dataset(batch: List[Example]) -> xr.Dataset:
-    """Concat all the individual fields in an Example into a single Dataset.
-
-    Args:
-      batch: List of Example objects, which together constitute a single batch.
-    """
-    datasets = []
-    for i, example in enumerate(batch):
-        try:
-            individual_datasets = []
-            example_dim = {"example": np.array([i], dtype=np.int32)}
-            for name in ["sat_data", "nwp"]:
-                ds = example[name].to_dataset(name=name)
-                short_name = name.replace("_data", "")
-                if name == "nwp":
-                    ds = ds.rename({"target_time": "time"})
-                for dim in ["time", "x", "y"]:
-                    ds = coord_to_range(ds, dim, prefix=short_name)
-                ds = ds.rename(
-                    {
-                        "variable": f"{short_name}_variable",
-                        "x": f"{short_name}_x",
-                        "y": f"{short_name}_y",
-                    }
-                )
-                individual_datasets.append(ds)
-
-            # Datetime features
-            for name in DATETIME_FEATURE_NAMES:
-                ds = example[name].rename(name).to_xarray().to_dataset().rename({"index": "time"})
-                ds = coord_to_range(ds, "time", prefix=None)
-                individual_datasets.append(ds)
-
-            # PV
-            one_dateset = xr.DataArray(example["pv_yield"], dims=["time", "pv_system"])
-            one_dateset = one_dateset.to_dataset(name="pv_yield")
-            n_pv_systems = len(example["pv_system_id"])
-
-            # GSP
-            n_gsp_systems = len(example[GSP_SYSTEM_ID])
-            one_dateset['gsp_yield'] = xr.DataArray(example[GSP_YIELD], dims=["time_30", "gsp_system"])
-
-            # This will expand all dataarrays to have an 'example' dim.
-            # 0D
-            for name in ["x_meters_center", "y_meters_center"]:
-                try:
-                    one_dateset[name] = xr.DataArray([example[name]], coords=example_dim, dims=["example"])
-                except Exception as e:
-                    _LOG.error(f'Could not make pv_yield data for {name} with example_dim={example_dim}')
-                    if name not in example.keys():
-                        _LOG.error(f'{name} not in data keys: {example.keys()}')
-                    _LOG.error(e)
-                    raise Exception
-
-            # 1D
-            for name in ["pv_system_id", "pv_system_row_number", "pv_system_x_coords", "pv_system_y_coords"]:
-                one_dateset[name] = xr.DataArray(
-                    example[name][None, :],
-                    coords={**example_dim, **{"pv_system": np.arange(n_pv_systems, dtype=np.int32)}},
-                    dims=["example", "pv_system"],
-                )
-
-            # GSP
-            for name in [GSP_SYSTEM_ID, GSP_SYSTEM_X_COORDS, GSP_SYSTEM_Y_COORDS]:
-                try:
-                    one_dateset[name] = xr.DataArray(
-                        example[name][None, :],
-                        coords={**example_dim, **{"gsp_system": np.arange(n_gsp_systems, dtype=np.int32)}},
-                        dims=["example", "gsp_system"],
-                    )
-                except Exception as e:
-                    _LOG.debug(f'Could not add {name} to dataset. {example[name].shape}')
-                    _LOG.error(e)
-                    raise e
-
-            individual_datasets.append(one_dateset)
-
-            # Merge
-            merged_ds = xr.merge(individual_datasets)
-            datasets.append(merged_ds)
-
-        except Exception as e:
-            print(e)
-            _LOG.error(e)
-            raise Exception
-
-    return xr.concat(datasets, dim="example")
-
-
-def coord_to_range(da: xr.DataArray, dim: str, prefix: Optional[str], dtype=np.int32) -> xr.DataArray:
-    # TODO: Actually, I think this is over-complicated?  I think we can
-    # just strip off the 'coord' from the dimension.
-    coord = da[dim]
-    da[dim] = np.arange(len(coord), dtype=dtype)
-    if prefix is not None:
-        da[f"{prefix}_{dim}_coords"] = xr.DataArray(coord, coords=[da[dim]], dims=[dim])
-    return da
