+++ tests/test_datamodule.py
@@ -10,10 +10,9 @@ import pytest
 import nowcasting_dataset
 from nowcasting_dataset import datamodule
 from nowcasting_dataset.config.load import load_yaml_configuration
-from nowcasting_dataset.data_sources.nwp_data_source import NWP_VARIABLE_NAMES
 from nowcasting_dataset.datamodule import NowcastingDataModule
 from nowcasting_dataset.example import validate_example
-from nowcasting_dataset.dataset import batch_to_dataset
+from nowcasting_dataset.dataset.batch import batch_to_dataset
 from nowcasting_dataset.example import Example
 
 logging.basicConfig(format='%(asctime)s %(levelname)s %(pathname)s %(lineno)d %(message)s')
